+++ src/input/input_hls.cpp
@@ -879,7 +879,7 @@ namespace Mist{
     // Write packet ID mappings
     JSON::Value thisMappings;
     JSON::Value thisMappingsR;
-    for (std::map<size_t, uint64_t>::iterator pidIt = pidMappingR.begin();
+    for (std::map<uint64_t, uint64_t>::iterator pidIt = pidMappingR.begin();
          pidIt != pidMappingR.end(); pidIt++){
       thisMappingsR.append(pidIt->second);
     }
